+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added AcceptNullParts to the ConcatenateStringConverter to allow null part be part of the string.
 - Added DelegateConverter that provides a delegate to convert a single value or a list of values.
 - Added DelegateToBooleanConverter that provides a delegate to convert a single value or a list of values to boolean.
+- Added DelegateToVisibilityConverter that provides a delegate to convert a single value or a list of values to Visibility.
 ### Removed
 - Removed InverseBooleanConverter, thats replaced by BooleanToBooleanConverter.
 - Removed MultiBooleanToBooleanConverter, thats integrated into the BooleanToBooleanConverter.